+++ back/game-app/src/game/game.service.ts
@@ -20,11 +20,11 @@ import { GameCore } from './core/game.core';
 import { CheckDisconnectByReconnectionDto } from './dto/check-disconnect-by-reconnection.dto';
 import { PlayerAction } from './player/enums/player-action.enum';
 import { DisconnectGameDto } from './dto/disconnect-game-dto';
-import { Player } from './player/enums/player';
 import { PlayerStatus } from './player/enums/player-status.enum';
 import { PlayerNumber } from './player/enums/player-number.enum';
 import { MainService } from '../main/main.service';
 import { EndGameDto } from './dto/end-game.dto';
+import { Player } from './player/player';
 
 @Injectable()
 export class GameService {
